backend/event_collector/requirements.txt
anyio==4.6.2.post1
click==8.1.7
fastapi==0.93.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.10
pydantic==1.10.18
pytest==8.3.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.25.0
typing_extensions==4.12.2
uvicorn==0.20.0
